Move prettystate to libnmstate
