+++ libnmstate/prettystate.py
